Added support for coloured log messages in Chrome
When debugging Stativus in the browser there can be a bit stream of messages and it can be hard to pick out the important bits.  This commit simply makes ENTER messages green, EXIT messages red and ACTION messages purple (in Chrome) which makes debugging in the browser much easier on the eyes.